cherry pick pr 13 to rnw 52

+++ ReactWindows/ChakraBridge/NativeJavaScriptExecutor.cpp
@@ -39,7 +39,7 @@ ChakraStringResult NativeJavaScriptExecutor::GetGlobalVariable(String^ variableN
     size_t bufLen;
     IfFailRetNullPtr(JsStringToPointer(globalVariableJson, &szBuf, &bufLen));
 
-    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, bufLen) };
+    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, (unsigned int)bufLen) };
     return finalResult;
 }
 
@@ -98,7 +98,7 @@ ChakraStringResult NativeJavaScriptExecutor::CallFunctionAndReturnFlushedQueue(S
     size_t bufLen;
     IfFailRetNullPtr(JsStringToPointer(stringifiedResult, &szBuf, &bufLen));
 
-    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, bufLen) };
+    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, (unsigned int)bufLen) };
     return finalResult;
 }
 
@@ -136,7 +136,7 @@ ChakraStringResult NativeJavaScriptExecutor::InvokeCallbackAndReturnFlushedQueue
     size_t bufLen;
     IfFailRetNullPtr(JsStringToPointer(stringifiedResult, &szBuf, &bufLen));
 
-    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, bufLen) };
+    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, (unsigned int)bufLen) };
     return finalResult;
 }
 
@@ -165,6 +165,6 @@ ChakraStringResult NativeJavaScriptExecutor::FlushedQueue()
     size_t bufLen;
     IfFailRetNullPtr(JsStringToPointer(stringifiedResult, &szBuf, &bufLen));
 
-    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, bufLen) };
+    ChakraStringResult finalResult = { JsNoError, ref new String(szBuf, (unsigned int)bufLen) };
     return finalResult;
 }
